# Conflicts:
#	.travis.yml
#	src/com/sheepit/client/Client.java
#	src/com/sheepit/client/Configuration.java
#	src/com/sheepit/client/Error.java
#	src/com/sheepit/client/Job.java
#	src/com/sheepit/client/Server.java
#	src/com/sheepit/client/SettingsLoader.java
#	src/com/sheepit/client/standalone/GuiSwing.java
#	src/com/sheepit/client/standalone/GuiTextOneLine.java
#	src/com/sheepit/client/standalone/Worker.java
#	src/com/sheepit/client/standalone/swing/activity/Settings.java
#	src/com/sheepit/client/standalone/swing/activity/Working.java

+++ src/com/sheepit/client/Job.java
@@ -36,6 +36,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.TimeZone;
+import java.util.Timer;
+import java.util.TimerTask;
 
 import com.sheepit.client.Configuration.ComputeType;
 import com.sheepit.client.Error.Type;
@@ -224,12 +226,15 @@ public class Job {
 	public Error.Type render() {
 		gui.status("Rendering");
 		RenderProcess process = getProcessRender();
+		Timer timerOfMaxRenderTime = null;
 		String core_script = "";
 		if (getUseGPU() && config.getGPUDevice() != null && config.getComputeMethod() != ComputeType.CPU) {
 			core_script = "sheepit_set_compute_device(\"CUDA\", \"GPU\", \"" + config.getGPUDevice().getCudaName() + "\")\n";
+			gui.setComputeMethod("GPU");
 		}
 		else {
 			core_script = "sheepit_set_compute_device(\"NONE\", \"CPU\", \"CPU\")\n";
+			gui.setComputeMethod("CPU");
 		}
 		core_script += String.format("bpy.context.scene.render.tile_x = %1$d\nbpy.context.scene.render.tile_y = %1$d\n", getTileSize());
 		File script_file = null;
@@ -306,6 +311,24 @@ public class Job {
 			getProcessRender().setProcess(os.exec(command, new_env));
 			BufferedReader input = new BufferedReader(new InputStreamReader(getProcessRender().getProcess().getInputStream()));
 			
+			if (config.getMaxRenderTime() > 0) {
+				timerOfMaxRenderTime = new Timer();
+				timerOfMaxRenderTime.schedule(new TimerTask() {
+					@Override
+					public void run() {
+						RenderProcess process = getProcessRender();
+						if (process != null) {
+							long duration = (new Date().getTime() - process.getStartTime() ) / 1000; // in seconds
+							if (config.getMaxRenderTime() > 0 &&  duration > config.getMaxRenderTime()) {
+								log.debug("Killing render because process duration");
+								OS.getOS().kill(process.getProcess());
+								setAskForRendererKill(true);
+							}
+						}
+					}
+				}, config.getMaxRenderTime() * 1000 + 2000); // +2s to be sure the delay is over
+			}
+			
 			long last_update_status = 0;
 			log.debug("renderer output");
 			try {
@@ -354,6 +377,9 @@ public class Job {
 		
 		int exit_value = process.exitValue();
 		process.finish();
+		if (timerOfMaxRenderTime != null) {
+			timerOfMaxRenderTime.cancel();
+		}
 		
 		if (script_file != null) {
 			script_file.delete();
@@ -372,6 +398,13 @@ public class Job {
 		
 		if (getAskForRendererKill()) {
 			log.debug("Job::render been asked to end render");
+			
+			long duration = (new Date().getTime() - process.getStartTime() ) / 1000; // in seconds
+			if (config.getMaxRenderTime() > 0 && duration > config.getMaxRenderTime()) {
+				log.debug("Render killed because process duration (" + duration + "s) is over user setting (" + config.getMaxRenderTime() + "s)");
+				return Error.Type.RENDERER_KILLED_BY_USER_OVER_TIME;
+			}
+			
 			if (files.length != 0) {
 				new File(files[0].getAbsolutePath()).delete();
 			}
@@ -600,6 +633,17 @@ public class Job {
 			// Blender quit
 			// end of rendering
 			return Type.RENDERER_OUT_OF_VIDEO_MEMORY;
+		}
+		else if (line.indexOf("CUDA error: Invalid value in cuTexRefSetAddress(") != -1) {
+			// Fra:83 Mem:1201.77M (0.00M, Peak 1480.94M) | Time:00:59.30 | Mem:894.21M, Peak:894.21M | color 3, RenderLayer | Updating Mesh | Copying Strands to device
+			// Fra:83 Mem:1316.76M (0.00M, Peak 1480.94M) | Time:01:02.84 | Mem:1010.16M, Peak:1010.16M | color 3, RenderLayer | Cancel | CUDA error: Invalid value in cuTexRefSetAddress(NULL, texref, cuda_device_ptr(mem.device_pointer), size)
+			// Error: CUDA error: Invalid value in cuTexRefSetAddress(NULL, texref, cuda_device_ptr(mem.device_pointer), size)
+			// Fra:83 Mem:136.82M (0.00M, Peak 1480.94M) | Time:01:03.40 | Sce: color 3 Ve:0 Fa:0 La:0
+			// Blender quit
+			// CUDA error: Invalid value in cuTexRefSetAddress(NULL, texref, cuda_device_ptr(mem.device_pointer), size)
+			// Refer to the Cycles GPU rendering documentation for possible solutions:
+			// http://www.blender.org/manual/render/cycles/gpu_rendering.html
+			return Error.Type.RENDERER_OUT_OF_VIDEO_MEMORY;
 		}
 		else if (line.indexOf("CUDA device supported only with compute capability") != -1) {
 			// found bundled python: /tmp/xx/2.73/python
